@@ -68,47 +68,61 @@
     }
   },
   "dependencies": {
-    "body-parser": "1.15.2",
+    "body-parser": "1.19.0",
     "check-node-version": "4.0.1",
     "electron-updater": "2.8.9",
-    "express": "4.14.0",
-    "flag-icon-css": "3.2.1",
-    "http-shutdown": "1.1.0",
+    "express": "4.17.1",
+    "flag-icon-css": "3.4.5",
+    "http-shutdown": "1.2.1",
     "json-stable-stringify": "1.0.1",
-    "lodash": "4.17.11",
+    "lodash": "4.17.15",
     "node-forge": "0.6.46",
-    "pm2": "2.10.1",
-    "request": "2.79.0",
-    "request-stats": "2.0.1",
-    "semver": "5.3.0",
-    "sudo-prompt": "6.2.1",
-    "taskkill": "2.0.0",
-    "tasklist": "3.1.0",
+    "pm2": "4.1.2",
+    "request": "2.88.0",
+    "request-stats": "3.0.0",
+    "semver": "6.3.0",
+    "sudo-prompt": "9.0.0",
+    "taskkill": "3.1.0",
+    "tasklist": "4.0.1",
     "tough-cookie-file-store": "1.2.0",
-    "universal-analytics": "0.4.9",
-    "yargs": "6.5.0",
-    "opn": "5.1.0",
-    "prismjs": "1.8.4",
-    "classnames": "2.2.5",
+    "universal-analytics": "0.4.20",
+    "yargs": "14.2.0",
+    "open": "7.0.0",
+    "prismjs": "1.17.1",
+    "classnames": "2.2.6",
     "netmask": "1.0.6",
-    "uuid": "3.2.1",
-    "compression": "1.7.2",
-    "ws": "5.1.1",
+    "uuid": "3.3.3",
+    "compression": "1.7.4",
+    "ws": "7.2.0",
     "diacritic": "0.0.2",
-    "underscore": "1.8.3",
     "rimraf": "2.5.2",
     "log-buffer": "0.0.3",
-    "cheerio": "1.0.0-rc.2",
-    "prompt-sync": "4.1.6",
+    "cheerio": "1.0.0-rc.3",
+    "prompt-sync": "4.1.7",
     "lum_lokijs": "1.5.11",
     "lum_socksv5": "0.0.11",
-    "ps-list": "5.0.0",
+    "ps-list": "6.3.0",
     "download": "7.1.0",
     "extract-zip": "1.6.7",
-    "open": "6.2.0",
     "cookie": "0.4.0",
     "winston": "3.2.1",
-    "winston-daily-rotate-file": "4.1.0"
+    "winston-daily-rotate-file": "4.2.1",
+    "bootstrap": "3.4.1",
+    "codemirror": "5.49.2",
+    "es6-shim": "0.35.5",
+    "event-target-shim": "5.0.1",
+    "jquery": "3.4.1",
+    "moment": "2.24.0",
+    "react": "16.11.0",
+    "react-dom": "16.11.0",
+    "react-router-dom": "5.1.2",
+    "react-bootstrap": "0.31.0",
+    "react-bootstrap-typeahead": "3.4.7",
+    "react-waypoint": "9.0.2",
+    "react-select": "3.0.8",
+    "react-tooltip": "3.11.1",
+    "react-virtualized": "9.21.2",
+    "file-saver": "2.0.2"
   },
   "description": "A configurable local proxy for luminati.io",
   "devDependencies": {
