@@ -16,7 +16,7 @@ import {T} from './common/i18n.js';
 
 class Stats extends Pure_component {
     state = {stats: {}, request_stats: false, toggling: false};
-    componentWillMount(){
+    componentDidMount(){
         this.setdb_on('head.recent_stats', stats=>{
             if (stats)
                 this.setState({stats});
@@ -27,7 +27,7 @@ class Stats extends Pure_component {
         });
         this.setdb_on('head.save_settings', save_settings=>{
             this.save_settings = save_settings;
-            if (typeof this.saved_stats == 'boolean')
+            if (typeof this.saved_stats=='boolean')
             {
                 this.toggle_stats(this.saved_stats);
                 delete this.saved_stats;
