@@ -8,7 +8,6 @@ const etask = require('../util/etask.js');
 const csv = require('../util/csv.js');
 const qw = require('../util/string.js').qw;
 const file = require('../util/file.js');
-const date = require('../util/date.js');
 const zerr = require('../util/zerr.js');
 const lpm_file = require('../util/lpm_file.js');
 const zcountry = require('../util/country.js');
@@ -21,27 +20,17 @@ const E = module.exports;
 let data;
 let pending;
 
-const fetch = endpoint=>etask(function*(){
-    let res;
-    try {
-        const headers = {'user-agent': util.user_agent};
-        res = yield util.json(endpoint, {headers, timeout: 20*date.ms.SEC});
-    } catch(e){
-        res = {};
-    }
-    return res;
-});
-
 const get_current_carrier_version = domain=>etask(function*(){
-    const res = yield fetch(`http://client.${domain}/api/carriers/version`);
+    const res = yield util.fetch(
+        `http://client.${domain}/api/carriers/version`);
     if (res.statusCode==200)
         return res.body;
-    logger.warn('Unable to get current carriers version. Carriers will not be '
+    logger.warn('Unable to get current carriers version. carriers will not be '
     +'displayed');
 });
 
 const get_current_carrier_data = domain=>etask(function*(){
-    const res = yield fetch(`http://client.${domain}/api/carriers/asns`);
+    const res = yield util.fetch(`http://client.${domain}/api/carriers/asns`);
     if (res.statusCode==200)
         return res.body;
     logger.warn('Unable to get current carrier data. Carriers will not be '+
@@ -53,13 +42,13 @@ const get_asn_carrier_map = domain=>etask(function*(){
     let carrier2asn = {};
     try { carrier2asn = JSON.parse(file.read_e(carriers_path)); }
     catch(e){
-        if (e.code != 'ENOENT')
+        if (e.code!='ENOENT')
             logger.warn('Failed to parse carriers %s', zerr.e2s(e));
     }
     const current_carrier_version = yield get_current_carrier_version(domain);
     if (!current_carrier_version)
         return {};
-    if (carrier2asn.version != current_carrier_version)
+    if (carrier2asn.version!=current_carrier_version)
     {
         logger.notice('Fetching carriers data. New version %s available '
         +'(current is %s)', current_carrier_version, carrier2asn.version);
@@ -74,7 +63,7 @@ const get_asn_carrier_map = domain=>etask(function*(){
     {
         for (const asn of carrier2asn.data[carrier])
         {
-            if (asn2carrier[asn] && asn2carrier[asn] != carrier)
+            if (asn2carrier[asn] && asn2carrier[asn]!=carrier)
             {
                 console.log('conflict on asn %s, %s, %s', asn,
                     asn2carrier[asn], carrier);
@@ -88,52 +77,55 @@ const get_asn_carrier_map = domain=>etask(function*(){
 const prepare_data = (csv_data, vipdb_data, asn2carrier)=>{
     csv_data.shift();
     const cities = [], structured = {};
-    csv_data.forEach(d=>{
-        if (!d)
-            return;
-        const r_id = d[1] && d[1].toLowerCase();
-        const c_id = d[3] && d[3].toLowerCase();
-        const city_id = diacritic.clean(d[0]).toLowerCase()
-            .replace(/[^a-z0-9]/g, '');
-        if (!c_id)
-            return;
-        const city_vip = _.get(vipdb_data,
-            `locations.${c_id}.${r_id}.${city_id}`);
-        if (d[0] && city_vip)
-        {
-            cities.push({
-                city_name: d[0],
-                region_id: r_id,
-                country_id: c_id,
-                mob: city_vip._mob,
-            });
-        }
-        if (!structured[c_id] && vipdb_data.locations[c_id])
-        {
-            structured[c_id] = {
-                country_name: zcountry.code2label(c_id),
-                country_id: c_id,
-                mob: vipdb_data.locations[c_id]._mob,
-                regions: {},
-                asns: vipdb_data.asns[c_id],
-                carriers: [...new Set(Object.keys(vipdb_data.asns[c_id])
-                    .map(asn=>asn2carrier[asn]).filter(Boolean))],
-            };
-        }
-        const region_vip = _.get(vipdb_data, `locations.${c_id}.${r_id}`);
-        if (r_id && region_vip)
-        {
-            const regions = structured[c_id].regions;
-            if (!regions[r_id])
+    if (vipdb_data)
+    {
+        csv_data.forEach(d=>{
+            if (!d)
+                return;
+            const r_id = d[1] && d[1].toLowerCase();
+            const c_id = d[3] && d[3].toLowerCase();
+            const city_id = diacritic.clean(d[0]).toLowerCase()
+                .replace(/[^a-z0-9]/g, '');
+            if (!c_id)
+                return;
+            const city_vip = _.get(vipdb_data,
+                `locations.${c_id}.${r_id}.${city_id}`);
+            if (d[0] && city_vip)
             {
-                regions[r_id] = {
+                cities.push({
+                    city_name: d[0],
                     region_id: r_id,
-                    region_name: d[2],
-                    mob: region_vip._mob,
+                    country_id: c_id,
+                    mob: city_vip._mob,
+                });
+            }
+            if (!structured[c_id] && vipdb_data.locations[c_id])
+            {
+                structured[c_id] = {
+                    country_name: zcountry.code2label(c_id),
+                    country_id: c_id,
+                    mob: vipdb_data.locations[c_id]._mob,
+                    regions: {},
+                    asns: vipdb_data.asns[c_id],
+                    carriers: [...new Set(Object.keys(vipdb_data.asns[c_id])
+                        .map(asn=>asn2carrier[asn]).filter(Boolean))],
                 };
             }
-        }
-    });
+            const region_vip = _.get(vipdb_data, `locations.${c_id}.${r_id}`);
+            if (r_id && region_vip)
+            {
+                const regions = structured[c_id].regions;
+                if (!regions[r_id])
+                {
+                    regions[r_id] = {
+                        region_id: r_id,
+                        region_name: d[2],
+                        mob: region_vip._mob,
+                    };
+                }
+            }
+        });
+    }
     const regions = _.fromPairs(_.toPairs(structured).map(p=>[
         p[0],
         _.values(p[1].regions).sort(
@@ -144,7 +136,8 @@ const prepare_data = (csv_data, vipdb_data, asn2carrier)=>{
     const countries = _.values(structured)
         .map(c=>_.pick(c, qw`country_id country_name mob`))
         .sort((m, n)=>(m.country_name||'').localeCompare(n.country_name));
-    return {cities, regions, countries, asns: vipdb_data.asns, carriers};
+    const asns = vipdb_data && vipdb_data.asns || {};
+    return {cities, regions, countries, asns, carriers};
 };
 
 E.ensure_data = www_domain=>etask(function*(){
@@ -156,8 +149,10 @@ E.ensure_data = www_domain=>etask(function*(){
         return yield this.wait();
     }
     pending = [];
-    const [vipdb_data, asn2carrier] = yield etask.all(
-        [locations.all_locations(), get_asn_carrier_map(www_domain)]);
+    const [vipdb_data, asn2carrier] = yield etask.all([
+        locations.all_locations(www_domain),
+        get_asn_carrier_map(www_domain),
+    ]);
     const csv_data = [];
     const cities_csv = path.resolve(__dirname, '../db/cities.csv');
     readline.createInterface({input: fs.createReadStream(cities_csv)})
