@@ -44,7 +44,7 @@ class Config {
             return v===def;
         }));
         proxies = proxies.map(p=>_.omitBy(p,
-            (v, k)=>k != 'zone' && v===this.defaults[k]));
+            (v, k)=>k!='zone' && v===this.defaults[k]));
         _defaults = _(_defaults)
         .pick(lpm_config.default_fields.filter(f=>f!='config'))
         .omitBy((v, k)=>{
