@@ -16,7 +16,7 @@ const lpm_headers = qw`x-hola-context proxy-authorization x-hola-agent
     x-lpm-ip`;
 const req_util = {
     is_ssl: req=>req.socket instanceof tls.TLSSocket,
-    is_connect: req=>req.method == 'CONNECT',
+    is_connect: req=>req.method=='CONNECT',
     full_url: req=>{
         if (!req_util.is_ssl(req))
             return req.url;
@@ -54,7 +54,7 @@ module.exports = class Context {
     }
     static init_req_ctx(req, res, serv, opt){
         const port = (opt||{}).listen_port;
-        if (req.ctx && req.ctx.port && port && req.ctx.port != port)
+        if (req.ctx && req.ctx.port && port && req.ctx.port!=port)
             req.ctx.port = port;
         req.ctx = req.ctx || new Context(req, res, serv, opt);
         req.ctx.request_start();
@@ -81,7 +81,7 @@ module.exports = class Context {
         this.is_ssl = req_util.is_ssl(this.req);
     }
     init_stats(){
-        if (this.retry > 0)
+        if (this.retry>0)
             return;
         const sp = etask(function*req_stats(){
             yield this.wait();
@@ -154,7 +154,7 @@ module.exports = class Context {
     }
     complete_req(){
         this.ref--;
-        if (this.ref > 0)
+        if (this.ref>0)
             return;
         this.sp.return();
         delete this.req;
