@@ -33,7 +33,7 @@ const levels = {
 winston.addColors({notice: 'cyan'});
 
 const create_logger = ()=>{
-    const log = yargs(process.argv.slice(2).map(String)).argv.log;
+    const log = yargs.parse(process.argv.slice(2).map(String)).log;
     const logger = winston.createLogger(get_opts(log));
     return logger;
 };
