@@ -160,5 +160,6 @@ E.prototype.stats_group_by = function(group_by, count=20){
 };
 
 E.prototype.stop = function(){
-    this.loki.autosaveDisable();
+    if (this.loki)
+        this.loki.autosaveDisable();
 };
