@@ -176,11 +176,10 @@ E.prototype.handle_usage = function(data){
         this.logs_process(data);
 };
 
-E.prototype.handle_usage_abort = function(data){
+E.prototype.handle_usage_abort = function(uuid){
     if (!this.wss)
         return;
-    if (data.context!='SESSION KEEP ALIVE')
-        this.logs_process(data);
+    this.wss.broadcast(uuid, 'har_viewer_abort');
 };
 
 E.prototype.handle_usage_start = function(data){
@@ -285,7 +284,8 @@ E.prototype.har = function(entries){
         entries: entries.map(entry=>{
             const req = JSON.parse(entry.request_headers||'{}');
             const res = JSON.parse(entry.response_headers||'{}');
-            const timeline = JSON.parse(entry.timeline||'[{}]');
+            const tl = (JSON.parse(entry.timeline||null)||[{}])[0];
+            const timeline = JSON.parse(entry.timeline||null)||[{}];
             entry.request_body = entry.request_body||'';
             const start = timeline[0].create;
             return {
@@ -313,7 +313,7 @@ E.prototype.har = function(entries){
                     remote_address: entry.remote_address,
                     rules: entry.rules,
                 },
-                startedDateTime: new Date(start||0).toISOString(),
+                startedDateTime: new Date(tl.create||0).toISOString(),
                 time: timeline.slice(-1)[0].end-start,
                 request: {
                     method: entry.method,
@@ -1072,7 +1072,7 @@ function*mgr_proxy_port_check(_this, port, duplicate, old_port, old_duplicate){
     const ports = [];
     for (let p = start; p <= end; p++)
     {
-        if (old_port && old_port <= p && p <= old_end)
+        if (old_port && old_port<=p && p<=old_end)
             continue;
         if (p==_this.argv.www)
             return p+' in use by the UI/API';
@@ -1322,7 +1322,7 @@ E.prototype.proxy_tester_api = function(req, res){
     const opt = Object.assign(_.pick(req.body, qw`url headers body`), {
         followRedirect: false,
     });
-    if (opt.body && typeof opt.body != 'string')
+    if (opt.body && typeof opt.body!='string')
         opt.body = JSON.stringify(opt.body);
     const password = proxy.config.password;
     const user = 'tool-proxy_tester';
@@ -1435,11 +1435,19 @@ E.prototype.filtered_get = function(req){
         query.port = {'$between': [req.query.port_from, req.query.port_to]};
     if (req.query.search)
         query.url = {'$regex': RegExp(req.query.search)};
-    ['port', 'content_type', 'status_code', 'protocol'].forEach(param=>{
+    let status_code;
+    if (status_code = req.query.status_code)
+    {
+        if (/^\d\*\*$/.test(status_code))
+            query.status_code = {'$regex': RegExp(`^${status_code[0]}`)};
+        else
+            query.status_code = +status_code;
+    }
+    ['port', 'content_type', 'protocol'].forEach(param=>{
         let val;
         if (val = req.query[param])
         {
-            if (param=='port' || param=='status_code')
+            if (param=='port')
                 val = +val;
             query[param] = val;
         }
@@ -2070,7 +2078,7 @@ function*mgr_create_web_interface(_this){
         util_lib.find_iface(_this.argv.iface)||'0.0.0.0']);
     const port = server.address().port;
     let address = server.address().address;
-    if (address == '0.0.0.0')
+    if (address=='0.0.0.0')
         address = '127.0.0.1';
     server.url = `http://${address}:${port}`;
     swagger.host = `${address}:${port}`;
@@ -2133,20 +2141,20 @@ E.prototype.has_created_proxy_port = function(){
 E.prototype.first_lpm_action = etask._fn(
 function*mgr_first_lpm_action(_this, action, ua, no_retry){
     if (!_this.logged_in || _this.first_actions.sent[action] ||
-        _this.first_actions.sending[action] || action == 'create_proxy_port' &&
+        _this.first_actions.sending[action] || action=='create_proxy_port' &&
         _this.first_actions.sent.create_proxy_port_def)
     {
         return;
     }
     _this.first_actions.sending[action] = 1;
-    const pending = _this.first_actions.pending.find(d=>d.action == action);
+    const pending = _this.first_actions.pending.find(d=>d.action==action);
     if (pending)
         ua = pending.ua;
     const proxy_port_not_sent =
         !_this.first_actions.sent.create_proxy_port &&
         !_this.first_actions.pending.find(
             d=>d.action.startsWith('create_proxy_port'));
-    if (action == 'send_request' && proxy_port_not_sent)
+    if (action=='send_request' && proxy_port_not_sent)
     {
         if (_this.has_created_proxy_port())
             yield _this.first_lpm_action('create_proxy_port', ua);
@@ -2168,12 +2176,12 @@ function*mgr_first_lpm_action(_this, action, ua, no_retry){
         });
         _this.first_actions.sent[action] = 1;
         _this.first_actions.pending =
-            _this.first_actions.pending.filter(d=>d.action != action);
+            _this.first_actions.pending.filter(d=>d.action!=action);
         if (!no_retry)
             yield _this.retry_failed_first_actions();
     } catch(e){
         if (!_this.first_actions.sent[action] &&
-            !_this.first_actions.pending.find(d=>d.action == action))
+            !_this.first_actions.pending.find(d=>d.action==action))
         {
             _this.first_actions.pending.push({action, ua, msg: e.message});
         }
@@ -2208,7 +2216,7 @@ function*verify_two_token_api(_this, req, res){
     });
     if ([200, 204].includes(response.statusCode))
         return res.sendStatus(200);
-    if (response.body == 'token expired')
+    if (response.body=='token expired')
         yield _this.send_two_step_email();
     res.status(response.statusCode).send(response.body);
 });
@@ -2339,7 +2347,7 @@ function*mgr_login_user(_this, token, username, password, customer){
             jar,
             headers,
         }]);
-        if (login_resp.statusCode != 200)
+        if (login_resp.statusCode!=200)
             return {error: {message: login_resp.body}};
         _this.luminati_jar = {jar, username, password};
         try {
@@ -2599,7 +2607,7 @@ function*mgr_run_cpu_usage_monitoring(_this){
     while (1)
     {
         const usage = Math.round(zos.cpu_usage().all*100);
-        const level = usage >= 98 ? 'error' : usage >= 75 ? 'warn' : null;
+        const level = usage>=98 ? 'error' : usage>=75 ? 'warn' : null;
         if (_this.wss)
             _this.wss.broadcast({usage, level}, 'cpu_usage');
         if (level)
