@@ -43,7 +43,7 @@ class Rules {
             const rule = this.rules[i];
             if (rule.type!='before_send' && rule.type!='timeout')
                 continue;
-            if (!rule.trigger.test(opt, req.ctx))
+            if (!rule.trigger.test(opt))
                 continue;
             if (rule.type=='timeout' && opt.timeout)
             {
@@ -69,7 +69,7 @@ class Rules {
             const opt = {url: req.ctx.url, status, time_passed, _res};
             if (rule.type!='after_hdr')
                 continue;
-            if (!rule.trigger.test(opt, req.ctx) || req.ctx.skip_rule(rule))
+            if (!rule.trigger.test(opt) || req.ctx.skip_rule(rule))
                 continue;
             this.check_req_time_range(req, rule);
             if (this.action(req, res, head, rule, opt))
@@ -96,7 +96,7 @@ class Rules {
                 _res};
             if (rule.type!='after_body')
                 continue;
-            if (!rule.trigger.test(opt, req.ctx) || req.ctx.skip_rule(rule))
+            if (!rule.trigger.test(opt) || req.ctx.skip_rule(rule))
                 continue;
             this.check_req_time_range(req, rule);
             let res_action;
@@ -168,7 +168,7 @@ class Rules {
         if (!zurl.is_valid_url(opt.url))
             return;
         opt.method = opt.method || 'GET';
-        const has_payload = opt.method != 'GET' && !!opt.payload;
+        const has_payload = opt.method!='GET' && !!opt.payload;
         const url = /^https?:\/\//.test(opt.url) ? opt.url : 'http://'+opt.url;
         const req_opt = {url, method: opt.method};
         if (has_payload)
@@ -273,17 +273,20 @@ const E = module.exports = Rules;
 
 class Trigger {
     constructor(r){
-        this.code = r.trigger_code;
+        const code = r.trigger_code+'\nreturn trigger;';
+        try { this.func = new Function(code)(); }
+        catch(e) { logger.warn('trigger: invalid function'); }
     }
-    test(opt, ctx){
-        opt = opt||{};
-        const code = this.code+'\nreturn trigger;';
-        try {
-            const func = new Function(code)();
-            return func(opt);
-        } catch(e){ logger.warn('trigger: function failed'); }
+    test(opt){
+        if (!this.func)
+        {
+            logger.warn('trigger: invalid function');
+            return false;
+        }
+        try { return this.func(opt||{}); }
+        catch(e){ logger.warn('trigger: function failed'); }
         return false;
     }
 }
 
-E.Trigger = Trigger;
+E.t = {Trigger};
