@@ -476,10 +476,7 @@ E.prototype.complete_req = function(err, req, res){
             this.send_error(req, res, err, 'lpm');
         this.active--;
         if (!this.active)
-        {
-            this.emit('idle', true);
-            return;
-        }
+            return this.emit('idle', true);
         if (this.throttle_queue.length)
             this.throttle_queue.shift().continue();
     } catch(e){
@@ -701,7 +698,7 @@ E.prototype.lpm_request = etask._fn(
 function*lpm_request(_this, req, res, head){
     const headers = Object.assign({}, req.headers);
     const ua = headers['user-agent'];
-    const is_user_req = ()=>ua != user_agent && !headers['x-lpm-keep-alive'];
+    const is_user_req = ()=>ua!=user_agent && !headers['x-lpm-keep-alive'];
     if (is_user_req())
         _this.emit('first_lpm_action', {action: 'send_request', ua});
     const ctx = Context.init_req_ctx(req, res, _this, _this.opt);
@@ -746,7 +743,7 @@ function*lpm_request(_this, req, res, head){
             _this.rules.post(req, res, head, resp);
         if (resp=='switched')
         {
-            _this.usage_abort(req);
+            _this.emit('usage_abort', req.ctx.uuid);
             yield this.wait();
         }
         if (resp instanceof Error)
@@ -784,7 +781,7 @@ function*lpm_request(_this, req, res, head){
 E.prototype.handle_custom_error = function(e, req, res, ctx){
     if (!is_custom_error(e))
         return;
-    if (e.message == 'Authentication failed')
+    if (e.message=='Authentication failed')
     {
         this.logger.info('%s %s 502 %s', req.method, ctx.url, e.message);
         write_http_reply(res, {
@@ -1417,21 +1414,7 @@ E.prototype.abort_proxy_req = function(req, proxy, task){
     if (task)
         task.return('abort');
     // XXX krzysztof: fix retry
-    this.usage_abort(req);
-};
-
-E.prototype.usage_abort = function(req){
-    const data = {
-        uuid: req.ctx.uuid,
-        port: this.port,
-        url: req.url,
-        method: req.method,
-        headers: req.headers,
-        timestamp: Date.now(),
-        context: req.ctx.h_context,
-        status_code: 'canceled',
-    };
-    this.emit('usage_abort', data);
+    this.emit('usage_abort', req.ctx.uuid);
 };
 
 E.prototype.handle_proxy_error = function(req, res, proxy, task, head){
