@@ -28,7 +28,7 @@ class Timeline {
         return metric1-metric2;
     }
     get(name, idx, nofb){
-        if (typeof idx != 'number')
+        if (typeof idx!='number')
             idx = this.req_chain.length-1;
         if (!Array.isArray(name))
             name = [name];
