@@ -254,7 +254,7 @@ E.find_iface = iface=>{
 let timeouts = [];
 E.set_timeout = (cb, delay)=>{
     const timeout = setTimeout(()=>{
-        timeouts = timeouts.filter(t=>t != timeout);
+        timeouts = timeouts.filter(t=>t!=timeout);
         cb();
     }, delay);
     timeouts.push(timeout);
@@ -276,6 +276,17 @@ E.get_lpm_tasks = (opt={})=>etask(function*(){
     };
     return tasks.filter(t=>t.name.includes('node') && regex.test(t.cmd) &&
         compare_pid(t));
+});
+
+E.fetch = endpoint=>etask(function*(){
+    let res;
+    try {
+        const headers = {'user-agent': E.user_agent};
+        res = yield E.json(endpoint, {headers, timeout: 20*date.ms.SEC});
+    } catch(e){
+        res = {};
+    }
+    return res;
 });
 
 E.get_status_tasks_msg = tasks=>{
