@@ -2,6 +2,10 @@
 'use strict'; /*jslint node:true*/
 const webpack = require('webpack');
 const html_webpack_plugin = require('html-webpack-plugin');
+const mk_url_loader = mimetype=>[{
+    loader: 'url-loader',
+    options: {limit: '100000', fallback: 'file-loader', mimetype},
+}];
 module.exports = {
     context: `${__dirname}/src/pub`,
     entry: {
@@ -33,34 +37,32 @@ module.exports = {
     module: {
         rules: [
             {
-                test: /src[\\\/]pub[\\\/].+\.js$/,
+                test: /src[\\/]pub[\\/].+\.js$/,
                 exclude: /node_modules/,
                 use: ['babel-loader'],
             },
             {
-                test: /www[\\\/].+\.js$/,
+                test: /www[\\/].+\.js$/,
                 exclude: /node_modules/,
                 use: ['hutil-loader', 'babel-loader'],
             },
             {
-                test: /util[\\\/].+\.js$/,
+                test: /util[\\/].+\.js$/,
                 parser: {node: false, commonjs: false},
-                exclude: [/www[\\\/].+\.js$/, /node_modules/],
+                exclude: [/www[\\/].+\.js$/, /node_modules/],
                 use: ['hutil-loader'],
             },
             {test: /\.css$/, use: ['style-loader', 'css-loader']},
             {test: /\.less$/, use: ['style-loader', 'css-loader?-url',
                 'less-loader']},
-            {test: /\.eot(\?v=\d+.\d+.\d+)?$/,
-                use: 'url-loader?limit=100000&name=[name].[ext]'},
-            {test: /\.woff2?(\?v=[0-9]\.[0-9]\.[0-9])?$/, use: 'url-loader?'
-                    +'limit=100000&mimetype=application/font-woff&'
-                    +'name=[name].[ext]'},
-            {test: /\.[ot]tf(\?v=\d+.\d+.\d+)?$/, use: 'url-loader?'
-                    +'limit=100000&mimetype=application/octet-stream'},
-            {test: /\.svg(\?v=\d+.\d+.\d+)?$/, use: 'url-loader?'
-                +'limit=120000&mimetype=image/svg+xml&name=[name].[ext]'},
-            {test: /\.(jpe?g|png|ico|gif)$/, use: 'url-loader?limit=100000'},
+            {test: /\.eot(\?v=\d+.\d+.\d+)?$/, use: mk_url_loader()},
+            {test: /\.woff2?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+                use: mk_url_loader('application/font-woff')},
+            {test: /\.[ot]tf(\?v=\d+.\d+.\d+)?$/,
+                use: mk_url_loader('application/octet-stream')},
+            {test: /\.svg(\?v=\d+.\d+.\d+)?$/,
+                use: mk_url_loader('image/svg+xml')},
+            {test: /\.(jpe?g|png|ico|gif)$/, use: mk_url_loader()},
         ],
     },
     resolve: {
